requirements.txt
click==8.0.1
Flask==2.0.1
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.1.0
pymongo==3.12.0
pypng==0.20220715.0
python-dotenv==0.19.2
qrcode==7.4.2
typing_extensions==4.8.0
Werkzeug==2.0.1
